refs #2015 - better explanation when to use reverse_proxy = 1
The current behaviour as you observe in comment:2 is correct.

The reason why jhstatewide couldn't login with reverse_proxy=1 is likely
because he's using a non-https proxy (i.e., http to http); in which case, the
browser won't send back the secure-only cookie to Piwik over an http connection.

Setting reverse_proxy=0 and configuring the proxy headers should solve both the
wrong IPs and login issue.



git-svn-id: http://dev.piwik.org/svn/trunk@3734 59fd770c-687e-43c8-a1e3-f5a4ff64c105